Consolidate packages, add user configuration
This commit consolidates the `deuill-base-extras` and
`deuill-sway-extras` packages into the `deuill-sway` package, which
represents the complete desktop environment for the installed system;
the `deuill-base` package remains as a package used for handling basic,
system-wide configuration.

In addition to the above, user configuration will now be installed to
`/etc/skel`, to be copied to home directories for any new user, or to be
used with GNU Stow or similar systems.